+++ src/main/java/ladder/exception/generator/ContinuousPathException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.generator;
 
 public class ContinuousPathException extends RuntimeException {
     private static final String MESSAGE = "연속된 사다리 발판이 존재합니다.";
