+++ src/main/java/ladder/exception/generator/InvalidStepWidthException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.generator;
 
 public class InvalidStepWidthException extends RuntimeException {
     private static final String MESSAGE = "생성된 사다리 너비가 올바르지 않습니다.";
