+++ src/main/java/ladder/exception/ladder/InvalidHeightNumberException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.ladder;
 
 public class InvalidHeightNumberException extends IllegalArgumentException {
     private static final String MESSAGE = "사다리 높이는 숫자로 입력해주세요.";
