+++ src/main/java/ladder/exception/participant/DelimiterBoundaryException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.participant;
 
 public class DelimiterBoundaryException extends IllegalArgumentException {
     private static final String MESSAGE = "구분자는 양 끝에 입력할 수 없습니다.";
