+++ src/main/java/ladder/exception/participant/DuplicatedNamesException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.participant;
 
 public class DuplicatedNamesException extends IllegalArgumentException {
     private static final String MESSAGE = "중복된 이름이 입력되었습니다.";
