+++ src/main/java/ladder/exception/participant/InvalidNameLengthException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.participant;
 
 public class InvalidNameLengthException extends IllegalArgumentException {
     private static final String MESSAGE = "이름은 1에서 5자 사이로 입력해 주세요";
