+++ src/main/java/ladder/exception/participant/InvalidParticipantsCountException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.participant;
 
 public class InvalidParticipantsCountException extends IllegalArgumentException {
     private static final String MESSAGE = "참가자 수는 2명 이상입니다.";
