+++ src/main/java/ladder/exception/participant/NonAlphabeticNameException.java
@@ -1,4 +1,4 @@
-package ladder.exception;
+package ladder.exception.participant;
 
 public class NonAlphabeticNameException extends IllegalArgumentException {
     private static final String MESSAGE = "이름은 영어로 입력해 주세요";
