+++ src/test/java/ladder/domain/ladder/HeightTest.java
@@ -1,6 +1,6 @@
 package ladder.domain.ladder;
 
-import ladder.exception.InvalidHeightRangeException;
+import ladder.exception.ladder.InvalidHeightRangeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
