+++ src/test/java/ladder/domain/participant/ParticipantsTest.java
@@ -1,6 +1,6 @@
 package ladder.domain.participant;
 
-import ladder.exception.InvalidParticipantsCountException;
+import ladder.exception.participant.InvalidParticipantsCountException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
